Relevant BUGIDs:
Purpose of commit: bugfix

Commit summary:
---------------

2006-06-17  Thorsten Kukuk  <kukuk@thkukuk.de>

        * examples/vpass.c: UID is unsigned on Linux.
        * modules/pam_exec/pam_exec.c: Likewise.
        * modules/pam_unix/pam_unix_acct.c: Likewise.
        * modules/pam_unix/pam_unix_sess.c: Likewise.

        * modules/pam_succeed_if/pam_succeed_if.8.xml: Fix syntax error.
        * modules/pam_succeed_if/pam_succeed_if.8: Regenerated.
        * modules/pam_succeed_if/README: Regenerated.

+++ examples/vpass.c
@@ -33,7 +33,7 @@ int main(void)
     if (pw) {
 	user = pw->pw_name;
     } else {
-	fprintf(stderr, "Invalid userid: %d\n", uid);
+	fprintf(stderr, "Invalid userid: %lu\n", (unsigned long) uid);
 	exit(1);
     }
