+++ modules/pam_unix/pam_unix_acct.c
@@ -209,8 +209,8 @@ PAM_EXTERN int pam_sm_acct_mgmt(pam_handle_t * pamh, int flags,
 	D(("user = `%s'", uname));
 	if (retval != PAM_SUCCESS || uname == NULL) {
 		pam_syslog(pamh, LOG_ALERT,
-			 "could not identify user (from uid=%d)",
-			 getuid());
+			 "could not identify user (from uid=%lu)",
+			 (unsigned long int)getuid());
 		return PAM_USER_UNKNOWN;
 	}
